rename/move user file a bit
