+++ src/pages/api/changePassword.ts
@@ -2,13 +2,9 @@ import { Request } from "express";
 import { updateMutator } from "@vulcanjs/graphql";
 import { NextApiRequest, NextApiResponse } from "next";
 import { User, UserConnector, UserType } from "~/models/user";
-import {
-  StorableTokenConnector,
-  hashToken,
-} from "~/models/storableToken.server";
 
 import { contextFromReq } from "~/lib/api/context";
-import { checkPasswordForUser } from "~/lib/api/passport/account";
+import { checkPasswordForUser } from "~/lib/api/account";
 
 interface ChangePasswordBody {
   oldPassword: string;
